requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cnocr==2.2.2.2
cnstd==1.2.2
coloredlogs==15.0.1
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
Django==4.1.7
Flask==2.2.3
flatbuffers==23.3.3
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.3.0
humanfriendly==10.0
idna==3.4
image==1.5.33
importlib-metadata==6.1.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
lightning-utilities==0.8.0
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
multidict==6.0.4
numpy==1.24.2
onnx==1.13.1
onnxruntime==1.14.1
openai==0.27.2
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
Polygon3==3.0.9.1
protobuf==3.20.3
pyclipper==1.3.0.post4
pyocr==0.8.3
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
pytorch-lightning==2.0.1
pytz==2023.3
PyYAML==6.0
requests==2.28.2
retrying==1.3.4
scipy==1.10.1
seaborn==0.12.2
self==2020.12.3
shapely==2.0.1
six==1.16.0
sprint==0.1.7
sqlparse==0.4.3
sympy==1.11.1
torch==1.13.1
torchmetrics==0.11.4
torchvision==0.14.1
tqdm==4.65.0
typing-extensions==4.5.0
Unidecode==1.3.6
urllib3==1.26.15
Werkzeug==2.2.3
XLML==0.1.2
yarl==1.8.2
zipp==3.15.0
